Cast invoice-sequence column to BIGINT in postgres branch
The postgres path of next_invoice_number and
backfill_missing_booking_invoices was casting the parsed sequence as
INTEGER (INT4) and decoding the resulting MAX(seq) into Option<i64>
(INT8), which sqlx 0.8 rejects with:

  Invoice backfill failed: Database error: error occurred while decoding
  column 0: mismatched types; Rust type `core::option::Option<i64>`
  (as SQL type `INT8`) is not compatible with SQL type `INT4`

Cast directly to BIGINT so the column type matches the i64 binding.
SQLite's INTEGER affinity is already variable-width INT8 so the SQLite
branch needed no change.

+++ hotel-app-be/src/services/invoice_numbers.rs
@@ -39,10 +39,10 @@ pub async fn next_invoice_number(pool: &DbPool) -> Result<String, ApiError> {
     let max_seq: Option<i64> = sqlx::query_scalar(
         r#"
         SELECT MAX(seq) FROM (
-            SELECT CAST(SUBSTRING(invoice_number FROM 12) AS INTEGER) AS seq
+            SELECT CAST(SUBSTRING(invoice_number FROM 12) AS BIGINT) AS seq
             FROM invoices WHERE invoice_number LIKE $1
             UNION ALL
-            SELECT CAST(SUBSTRING(invoice_number FROM 12) AS INTEGER) AS seq
+            SELECT CAST(SUBSTRING(invoice_number FROM 12) AS BIGINT) AS seq
             FROM customer_ledgers WHERE invoice_number LIKE $1
         ) combined
         "#,
@@ -130,10 +130,10 @@ pub async fn backfill_missing_booking_invoices(pool: &DbPool) -> Result<usize, A
             let max_seq: Option<i64> = sqlx::query_scalar(
                 r#"
                 SELECT MAX(seq) FROM (
-                    SELECT CAST(SUBSTRING(invoice_number FROM 12) AS INTEGER) AS seq
+                    SELECT CAST(SUBSTRING(invoice_number FROM 12) AS BIGINT) AS seq
                     FROM invoices WHERE invoice_number LIKE $1
                     UNION ALL
-                    SELECT CAST(SUBSTRING(invoice_number FROM 12) AS INTEGER) AS seq
+                    SELECT CAST(SUBSTRING(invoice_number FROM 12) AS BIGINT) AS seq
                     FROM customer_ledgers WHERE invoice_number LIKE $1
                 ) combined
                 "#,
